New XmlInventory to serialize inventories new package for gadgets (gadgets, items with effect / performers) new event GadgetUseEvent NameTagEntity in entities.nms, invisible nameless bat!
Interfaces for gadgets and abstracted implementations.

Began working on Admin menu

Made entire new core for plugin writing, handling alot of the boiler plating when using commons to it's fullest. This includes a sort of "game" api where minigames, bukkit plugins, and games in general can be made so much easier.

extending MiniGame gives you a plugin base with all the tools required, all you do is register the gamestates and away she goes!

BukkitPlugin is rather bare bones, but, it's supposed to be.

CraftGame is a BukkitPlugin but with implementation of the GameCore

All interfaces are available so creating your own cores and implementing them is encouraged!

Began writing arena managment, player wrappers, and other things to ease development.