Commons 1.7 Refactor; Goal is to have all internal methods accessed via internal and object-specific methods, where-as the API (methods designed for public use) are left external.
Ontop of this, we'll have a new ItemSetManager, which can be instanced by plugins to create named 'kits' in a sense.

A saved-item manager is added for the /i save and /i load commands.

All methods pertaining to chat have been moved from Players to Chat.java

More to come.