Created the XMLTitle class to serialize titles Created the ServerInfo class to contain information about other servers on the network run with Commons Applied required patches to work with Spigots ProjectileSource class
Updated ItemMenu's with a load of bug fixes.
Updated MenuItem with a slight bug fix, and optimization.
Created the MenuBuilder to make menus with ease, now!

Created the ActionCooldown, used to set a string-based ation on cooldown.

Fixed the null-pointer-exception when loading sets from file on load of Commons.

Created the network package, inclusive is a Bungee class; More will be added as time flies.