Add support for both YML and XML based configuration files using a generic Configuration interface.
A data-option.txt file is generated based on (previous) configuration if they have it, or created when a new instance of Commons is initialized. Conversion between the two types of data should be handled with ease on startup.

Automatic type convertors for ItemStacks, Enchantments, Inventories, Sounds, Titles, Potions, Locations, ParticleEffects, Entities, and so forth will follow in the soon future.